+++ src/org/scourge/terrain/Terrain.java
@@ -185,7 +185,7 @@ public class Terrain implements NodeGenerator {
     }
 
     // called from the main thread
-    public void update(float tpf) {
+    public void update(double tpf) {
         // make an unsynchronized check (a small hack: pendingRegions.isEmpty() would have to be synchronized)
         if(checkPendingRegions > 0) {
             // add/remove regions in the main thread to avoid concurrent mod. exceptions
